requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
lxml==5.1.0
opencv-python==4.8.1.78
Pillow==10.1.0
